requirements.txt
aldryn-apphook-reload==0.3.0
aldryn-apphooks-config==0.5.2
aldryn-boilerplates==0.8.0
aldryn-common==1.0.4
aldryn-translation-tools==0.2.1
appnope==0.1.0
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
awscli==1.16.126
awsebcli==3.14.13
backcall==0.1.0
backports.csv==1.0.7
beautifulsoup4==4.7.1
bleach==2.1.4
blessed==1.15.0
boto==2.49.0
boto3==1.8.6
botocore==1.12.116
cached-property==1.4.3
cachetools==3.1.0
cement==2.8.2
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cmsplugin-filer==1.1.3
Collectfast==0.6.2
colorama==0.3.9
constantly==15.1.0
cryptography==2.2.2
cssselect==1.0.3
decorator==4.3.0
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.0.13
django-admin-sortable==2.1.7
django-admin-sortable2==0.6.21
django-aggregate-if==0.5
django-analytical==2.4.0
django-anymail==3.0
django-app-namespace-template-loader==0.4.1
django-appconf==1.0.2
django-appdata==0.2.1
django-autoslug==1.9.4
django-breadcrumbs==1.1.4a1
django-ckeditor==5.7.1
django-classy-tags==0.8.0
django-cms==3.6.0
django-common-helpers==0.9.2
django-contrib-comments==1.9.0
django-cookies-samesite==0.2.0
django-cors-headers==2.4.0
django-cors-middleware==1.3.1
django-countries==5.3.1
django-cron==0.5.1
django-csp==3.4
django-embed-video==1.2.0
django-filer==1.5.0
django-filter==2.0.0
django-formtools==2.1
django-google-maps==0.10.1
django-haystack==2.8.1
django-image-optimizer==1.0.0
django-import-export==1.2.0
django-ipware==2.1.0
django-jet==1.0.8
django-js-asset==1.1.0
django-language-field==0.0.3
django-languages==0.1
django-markdownx==2.0.28
django-mathfilters==0.4.0
django-memcached==0.1.2
django-meta==1.5.0
django-meta-mixin==0.3.0
django-modelcluster==4.4
django-money==0.14.3
django-mptt==0.9.1
django-newsletter==0.8b1
django-parler==1.9.2
django-polymorphic==2.0.2
django-robots==3.1.0
django-secure==1.0.1
django-sekizai==0.10.0
django-ses==0.8.10
django-settings-export==1.2.1
django-sortedm2m==1.5.0
django-spurl==0.6.5
django-sslify==0.2.7
django-standard-form==1.1.1
django-storages==1.7
django-tagging==0.4.6
django-taggit==0.23.0
django-taggit-autosuggest==0.3.7
django-taggit-templatetags==0.2.5
django-templatetag-sugar==1.0
django-treebeard==4.3
django-user-agents==0.3.2
django-xmlrpc==0.1.8
djangocms-admin-style==1.2.8
djangocms-apphook-setup==0.3.0
djangocms-attributes-field==1.0.0
djangocms-blog==1.0.0b1
djangocms-column==1.9.0
djangocms-comments==0.2.3
djangocms-file==2.3.0
djangocms-googlemap==1.1.1
djangocms-link==2.1.2
djangocms-picture==2.1.3
djangocms-snippet==2.1.0
djangocms-style==2.1.0
djangocms-text-ckeditor==3.8.0
djangocms-video==2.1.1
djangorestframework==3.8.2
djangotoolbox==1.8.0
docker==3.7.0
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
draftjs-exporter==2.1.5
easy-thumbnails==2.6
elasticsearch==6.3.1
elasticsearch-dsl==6.2.1
entrypoints==0.2.3
et-xmlfile==1.0.1
fastnumbers==2.1.1
feedparser==5.2.1
future==0.16.0
-e git://github.com/ixc/glamkit-url-breadcrumbs.git@60938bbddffb675714f5ae9c40f2c1edecbac507#egg=glamkit_url_breadcrumbs
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
html5lib==1.0.1
httplib2==0.11.3
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipaddress==1.0.22
ipykernel==4.8.2
ipython==7.0.1
ipython-genutils==0.2.0
ipywidgets==7.4.0
jdcal==1.4
jedi==0.12.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
ldif3==3.1.1
lxml==4.2.4
Markdown==2.6.11
MarkupSafe==1.0
mistune==0.8.3
mots-vides==2015.5.11
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.6.0
oauth2client==4.1.2
oauthlib==2.0.7
odfpy==1.4.0
openpyxl==2.6.2
pandocfilters==1.4.2
parsel==1.5.1
parso==0.3.1
pathspec==0.5.9
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
polib==1.1.0
prometheus-client==0.3.1
prompt-toolkit==1.0.15
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.6.0
py-moneyed==0.8.0
pyaes==1.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyDispatcher==2.0.5
Pygments==2.2.0
PyHamcrest==1.9.0
PyJWT==1.6.3
pyOpenSSL==19.0.0
pyparsing==2.2.0
python-card-me==0.9.3
python-dateutil==2.7.3
python-docx==0.8.7
python-dotenv==0.11.0
python-memcached==1.59
python-scrapyd-api==2.1.2
python-slugify==1.2.0
python3-openid==3.1.0
pytz==2018.4
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.1
queuelib==1.5.0
regex==2018.8.17
requests==2.20.1
requests-oauthlib==0.8.0
rsa==3.4.2
s3transfer==0.2.0
Scrapy==1.6.0
scrapy-djangoitem==1.1.1
scrapyd==1.2.0
semantic-version==2.5.0
Send2Trash==1.5.0
service-identity==18.1.0
simplegeneric==0.8.1
simplejson==3.16.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
sorl-thumbnail==12.4.1
soupsieve==1.9
sqlparse==0.3.0
surlex==0.2.0
tablib==0.13.0
tabulate==0.7.5
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
texttable==0.9.1
tinify==1.5.1
tornado==5.1
traitlets==4.3.2
Twisted==18.9.0
ua-parser==0.8.0
unicodecsv==0.14.1
Unidecode==0.4.21
uritemplate==3.0.0
urllib3==1.24.1
URLObject==2.4.3
user-agents==2.0
virtualenv==16.0.0
w3lib==1.20.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.51.0
widgetsnbextension==3.4.0
Willow==1.1
xlrd==1.2.0
xlwt==1.3.0
YURL==0.13
zope.interface==4.6.0
